[IMP] mail, web_unsplash, html_editor: improve check on attachment upload
Purpose
=======
Unsplash bypass a feature that when you go to an URL that doesn't have
a route, then it will read from `ir.attachment` an attachment of type
binary with an URL. This is not a security issue, because we can't
control the URL, but there's no reason to allow type that are not
images.

Add tests about access when uploading attachment for a message.

Check that portal can only upload images for the HTML editor,
and add a file size limit.

Task-4687269

closes odoo/odoo#228087

Related: odoo/upgrade#8483
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>