[FIX] html_builder: fix UI navigation order in add snippet dialog
Since the accessibility improvement in [9ae02d8], snippets previews
within the add snippet dialog are tabable.
It reveals an issue with the order of the columns depending on the
language of the interface: when the backend language and the frontend
language are not read in the same direction (RTL / LTR), the frontend
language is taken to display the previews (see [f9c77de]). This is the
right approach to show the snippets themselves, but the order of the
columns should be done according to the backend language, which is the
one that gives the instructions for the overall UI.

To reproduce:
- Set the admin's language to arabic
- Clear the cache and refresh your page
- Edit and open the add snippet dialog
- Navigate with Tab
=> The 1st focused snippet is in the wrong column compared with the rest
of the UI.

[9ae02d8]: https://github.com/odoo/odoo/commit/9ae02d80894d4043e49d8e2cad068f8018e6f113
[f9c77de]: https://github.com/odoo/odoo/commit/f9c77de84aa6ea705e5d3f129328fb2199103b9a

task-5109547

closes odoo/odoo#250250

X-original-commit: 05412af665f140923782aa703235c0b2e8150d47
Signed-off-by: Francois Georis (fge) <fge@odoo.com>
Signed-off-by: Robin Lejeune (role) <role@odoo.com>