[FIX] pos_event: restore missing registration answers
In [1], a refactoring of the POS framework was done, but in the
process some overlooked behaviors induced a complete loss of the
answers given in the registration process, as long as they do not
include an answer to a 'selection' question.

STEPS
=====

0. Create a new event with a few questions (no 'selection' one)
1. Add tickets so that you can buy one in the POS
2. Go in the POS (reload data if needed)
3. Buy a ticket and fill the attendee form (answer to all questions)
4. Continue the POS flow and pay for your ticket.
5. Go back to the back-end and event > attendees
6. Check your new attendee: no answers are linked to the record.

ISSUE
=====

TDLR - Two fields have the same comodel on event.registration.
This breaks the use of inverseMap as both are loaded in the pos.
As they are treated sequentially when connecting related records,
hidden side effects occur when both are not set.

Details:

This effect is due to two main issues. First, after [1], we compute
the whole model reference relations using processModelDefs method,
that accounts for inverse relations of o2m, m2o fields and provides
an inverseMap to be used in several places, notably on creation or
deletion of relational records from the POS.

However, pos_event presents a peculiar situation, as registrations
have two o2m fields, both loaded in _load_pos_data_fields, that are
linked to the model event.registration.answer: registration_answer_ids
and registration_answer_choice_ids, a subset of the first, just with
a domain to only include answers of 'selection' questions.

Meaning that the inverseMap will only use one of the two, in this case
registration_answer_choice_ids. In turn, this means that any update of
registration_id on the answer model will update that field, even if the
original update was done on registration_answer_ids...

Secondly, one could notice that this should still work, as the inverseMap
is used in all places, we should just update the records through the field
registration_answer_choice_ids.

So why does it not work and why are all answers removed? Because both fields
are loaded in _load_pos_data_fields, and in _sanitazeRawData we use getFields
to know which fields to update and to connect (for relations on the model),
both being returned.

This means that independantly of the create values for the registration
in addProductToOrder in pos_event, the 'framework' will still have two
fields to connect and will do so sequentially, one field at the time.

But in the _connect logic, if a field has no value given at creation,
then the 'CLEAR' command is used, as this would mean we remove the content
of the relational field. But in this case, this means that as both answer
fields try to update registration_id through registration_answer_choice_ids,
if no value is given at creation in that field, then we clear existing ones,
for instance those we just linked through the creation values in the first
field registration_answer_ids, as we basically empty registration_id on
those records. The answers will also be deleted of the indexedDB because of
the condition in databaseTable (no linked registration -> can be removed)

FIX
===

Remove registration_answer_choice_ids from the loaded data, as it is not
used anywhere except in the registration creation values. As it is a subset
of the other field registration_answer_ids, only keep that one instead.
Update the creation values to only use that field.

Note that this seems to highlight a limitation when it comes to having more
than one o2m field on a model loaded at the same time in POS.

[1] odoo/odoo@a80a39f2ad16baf474553574c79d55948f86c453

Task-4919080

closes odoo/odoo#249577

X-original-commit: eb0f31975537123fec4082b13468a23ad197307b
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Noé Antoine (nan) <nan@odoo.com>