[FIX] Profiler: Fix thread joining
Issue:
- Python throws cannot join current thread when using the
entry count with memory and traces async profilers.

Cause:
- The profiler has two ways to exit. One is through the context
which calls the end() through the parent thread. Another is
reaching the entry_count limit or whatever limit used. The exit
this way was called through the child thread of the collector.
This caused a deadlock as the thread was waiting for itself,
which caused python to through out an error "cannot join current
thread".

Solution:
- The solution for this issue is to set an have a single collector
run the end and set the done to true so no other collector would
end as well. Then, in the end the threads are joined to the ending
collector. The ending collector is not joined because it cannot
join itself. The event-loop of the ending thread would be broken
by setting the active field to false which would end the thread.

X-original-commit: 3b44d3b9ae3154eb63016d566b62554990ef8865
Part-of: odoo/odoo#249499
Signed-off-by: Xavier Dollé (xdo) <xdo@odoo.com>
Signed-off-by: Mohamed Mostafa Hamed (mmha) <mmha@odoo.com>