[FIX] Profiler: Thread sleep causes join hang
Issue:
- Issue was that the periodic profilers were waiting
the full amount of the sleep before ending.

Cause:
- This was caused because the join waits until the function
of the thread is done. Which means it would wait until the
sleep is done either way.

Solution:
- This was circumvented with the same way by using
an event wait. The event wait is broken as soon as the
event is set, but it waits a default amount of time before
releasing the wait.

X-original-commit: 5911d65878529358b9e5d686cf9305f421939735
Part-of: odoo/odoo#249499
Signed-off-by: Xavier Dollé (xdo) <xdo@odoo.com>
Signed-off-by: Mohamed Mostafa Hamed (mmha) <mmha@odoo.com>