[FIX] pos_sale: amount_unpaid wrong computation
When having a down payment that is reversed by a credit note,
the amount unpaid is wrongly computed.

This is because we take the sum of invoice lines price total, regardless
they come from invoice or credit note. Therefore we end up with negative
value.

Steps:

- Have a SO for 500
- Make a downpayment for 300, confirm
- Make a credit note for the downpayment invoice, confirm
-> SO's amount unpaid is -100, it should be 500. If you now settle the SO,
the amount unpaid will be -300 instead of 0.

opw-5175562

closes odoo/odoo#233248

Signed-off-by: David Monnom (moda) <moda@odoo.com>