[IMP] account_edi_ubl_cii: update factur-x/zugferd
Updating the FacturX format (France)/ ZUGFeRD format (Germany) to respect the new norms:
Factur-X 1.07.3 EXTENDED and ZUGFeRD 2.3.3 EXTENDED. Add the differentiation between these
two formats in the customer interface, even if they point to the same value in the code.
It clarifies things for the customer, things are called by their name.

Also, in Germany, for B2B invoices (peppol EAS = 9930), use ZUGFeRD, but for B2G invoices
(peppol EAS = 0204), use XRechnung. Adaptation of the default values in the partner form
according to this statement.

For French and German companies that are sending invoices to French, German or Belgian customers,
changed the default format of invoice sent to be compliant to PDF/A-3 norms.

task-5266286

closes odoo/odoo#237091

Signed-off-by: Laurent Smet (las) <las@odoo.com>