[FIX] l10n_ro: fixed romanian tax report
Romanian VAT has increased from 19% to 21% and from 5%/9% to 11%, some taxes
were added previously to the module but they were not reflected in the tax report,
also other taxes were missing in order for the report to replicate the current
up to date version issued by the romanian government, these taxes were added.
In addition, some taxes were no longer needed, instead of removing them, they were
set to inactive, as removing them would break the ability of some users to generate
the reports.
The order of the taxes after line 25 was changed since the total amount of lines
was reduced, so starting from line 26 there are inconsistenices between the line number
and the code in the tax report, since these are mainly for internal usage, there was no
need to update them, just the names and tax tags were updated to reflect the change
task-5411745

Part-of: odoo/odoo#241529
Related: odoo/enterprise#106127
Signed-off-by: Sven Führ (svfu) <svfu@odoo.com>